requirements.txt
discord.py
python-dotenv
discord.py[voice]
dotenv